+++ src/main/java/es/in2/verifier/config/ClientLoaderConfig.java
@@ -17,6 +17,7 @@ import org.springframework.security.oauth2.server.authorization.settings.ClientS
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Set;
 import java.util.UUID;
 
 @Configuration
@@ -24,6 +25,7 @@ import java.util.UUID;
 public class ClientLoaderConfig {
 
     private final TrustFrameworkService trustFrameworkService;
+    private final Set<String> allowedClientsOrigins;
 
     @Bean
     public RegisteredClientRepository registeredClientRepository() {
@@ -65,6 +67,11 @@ public class ClientLoaderConfig {
                 }
                 registeredClientBuilder.clientSettings(clientSettingsBuilder.build());
                 registeredClients.add(registeredClientBuilder.build());
+
+                // Add the client origin to the allowed clients origins
+                if (clientData.url() != null && !clientData.url().isBlank()) {
+                    allowedClientsOrigins.add(clientData.url());
+                }
             }
             return registeredClients;
         } catch (Exception e) {